+++ myMY/account.py
@@ -1,5 +1,5 @@
 """ account """
-from flask import Blueprint, render_template, url_for, flash, request, redirect
+from flask import Blueprint, render_template, url_for, flash, request, redirect, session
 from flask_login import current_user, login_required, fresh_login_required
 from flask_bcrypt import generate_password_hash, check_password_hash
 from .models import User
@@ -12,6 +12,7 @@ acc = Blueprint('acc', __name__)
 @login_required
 # account home
 def myAcc():
+    session['last'] = request.endpoint
     return render_template("account.html", user=current_user)
 
 
@@ -20,6 +21,7 @@ def myAcc():
 @fresh_login_required
 # account manager
 def accManager():
+    session['last'] = request.endpoint
     return render_template("account.html", user=current_user, title='Account Manager')
 
 
@@ -28,6 +30,7 @@ def accManager():
 @login_required
 # update password
 def updatePswd():
+    session['last'] = request.endpoint
     try:
         user = current_user
         curPswd = request.form.get("inputCurrentPassword")
@@ -63,6 +66,7 @@ def updatePswd():
 @login_required
 # update alias
 def updateAlias():
+    session['last'] = request.endpoint
     try:
         user = current_user
         curAlias = request.form.get("inputCurrentAlias")
@@ -99,6 +103,7 @@ def updateAlias():
 @login_required
 # delete account
 def delAcc():
+    session['last'] = request.endpoint
     try:
         user = current_user
         curPswd = request.form.get("inputCurrentPassword")