+++ myMY/idGen.py
@@ -5,6 +5,7 @@ from faker import *
 # uuidIns= UUID()
 fk = Faker()
 
+
 def gen1():
     """ gen uuid1 """
     return str(uuid1())