+++ CHANGELOG.md
@@ -36,4 +36,5 @@ Created on Dec 20, 2023
 - Beta Release: Attempt to fix overflow navbar on March 10, 2024 -> **0.622**
 - Beta Release: Navbar improvements on March 10, 2024 -> **0.624**
 - Beta Release: Navbar bugs fixed on March 10, 2024 -> **0.625**
-- Beta Release: New implementation on lookup transactions, overall improvements, and transaction modification implementation started on March 10, 2024 -> **0.7**
+- Beta Release: New implementation on lookup transactions, overall improvements, and re organized edit transaction on March 10, 2024 -> **0.7**
+- Beta Release: Overall improvements, and draft transaction modification on March 10, 2024 -> **0.72**